chore: extract signedArea2d from three local copies
generateCrenelatedWallManual, insetConvexPolygon and
SliceSatBoundingMeshRenderer each defined the same shoelace formula
over THREE.Vector2. SliceSat's variant early-returned 0 below three
points; the loop already yields 0 there, so dropping the guard is
behaviour-preserving.

generateUnionedLineRaftMesh keeps its own copy: it operates on Clipper
IntPoint (.X/.Y integer fields), not THREE.Vector2.
